% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost231.R
\name{cost231}
\alias{cost231}
\title{COST 231 path loss model}
\usage{
cost231(Data, freq)
}
\arguments{
\item{Data}{A dataframe with columns the distance from the transmiter and the measured power at that point.}

\item{freq}{the frequency in MHz of the transmition}
}
\value{
A list with the estimated Intercept of the model and the residuals of the propagation, i.e. Shadowing
\code{cost231} uses the following inputs:
}
\description{
Produces the estiamted power in each measured point based on the COST 231 path loss model
}
\seealso{
Other path loss models: \code{\link{erceg}},
  \code{\link{logDistance}}, \code{\link{okumura}},
  \code{\link{walfish}}
}
\concept{path loss models}
