% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erceg.R
\name{erceg}
\alias{erceg}
\title{Erceg HL/FM path loss model}
\usage{
erceg(Data, freq)
}
\arguments{
\item{Data}{A dataframe with columns the distance from the transmiter and the measured power at that point.}

\item{freq}{the frequency in MHz of the transmition}
}
\value{
A list with the estimated Intercept of the model and the residuals of the propagation, i.e. Shadowing
\code{erceg} uses the following inputs:
}
\description{
Produces the estiamted power in each measured point based on the Erceg HL/FM path loss model
}
\seealso{
Other path loss models: \code{\link{cost231}},
  \code{\link{logDistance}}, \code{\link{okumura}},
  \code{\link{walfish}}
}
\concept{path loss models}
