% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractShadowingInfo.R
\name{ExtractShadowingInfo}
\alias{ExtractShadowingInfo}
\title{Extract variogram and variogram fit information}
\usage{
ExtractShadowingInfo(Data, fittedModel, propModel)
}
\arguments{
\item{Data}{a character of the propagation model}

\item{fittedModel}{the variogram model to be fitted to data}

\item{propModel}{the propagation model to be used}
}
\value{
A list with the estimated Intercept of the model and the residuals of the propagation, i.e. Shadowing
\code{ExtractShadowingInfo} uses the following inputs:
}
\description{
Produces the estiamted power in each measured point based on the selected path loss model
}
\details{
The available propagation models are: "Log Distance", "Erceg HL/FM","Walfish-Ikegami","Okumura-Hata","COST 231"
}
\seealso{
Other fit: \code{\link{FitLinearPropModels}},
  \code{\link{fittedVariofit}}
}
\concept{fit}
