% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitLinearPropModels.R
\name{FitLinearPropModels}
\alias{FitLinearPropModels}
\title{Fit various path loss models}
\usage{
FitLinearPropModels(PropModel, Data, freq, typeOfArea)
}
\arguments{
\item{PropModel}{a character of the propagation model}

\item{Data}{A dataframe with columns the distance from the transmiter and the measured power at that point.}

\item{freq}{the frequency in MHz of the transmition}

\item{typeOfArea}{the type of area the measurement took place, i.e.: Urban or Suburban}
}
\value{
A list with the estimated Intercept of the model and the residuals of the propagation, i.e. Shadowing
\code{FitLinearPropModels} uses the following inputs:
}
\description{
Produces the estiamted power in each measured point based on the selected path loss model
}
\details{
The available propagation models are: "Log Distance", "Erceg HL/FM","Walfish-Ikegami","Okumura-Hata","COST 231"
}
\seealso{
Other fit: \code{\link{ExtractShadowingInfo}},
  \code{\link{fittedVariofit}}
}
\concept{fit}
