% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fittedVariofit.R
\name{fittedVariofit}
\alias{fittedVariofit}
\title{Fit the variogram depending on the max.dist based on the variofit of geoR}
\usage{
fittedVariofit(x, max.dist, scaled = FALSE, ...)
}
\arguments{
\item{x}{a variogram object/list}

\item{max.dist}{the maximum distance between points taken into consideration}
}
\description{
Produces a fit depending on the x object and the max. distance
}
\details{
The available covariance models are: "matern", "powered.exponential","cauchy",generalized cauchy ("gencauchy")
}
\seealso{
Other fit: \code{\link{ExtractShadowingInfo}},
  \code{\link{FitLinearPropModels}}
}
\concept{fit}
