% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logDistance.R
\name{logDistance}
\alias{logDistance}
\title{Log distance path loss model}
\usage{
logDistance(Data)
}
\arguments{
\item{Data}{A dataframe with columns the distance from the transmiter and the measured power at that point.}
}
\value{
A list with the estimated Intercept of the model and the residuals of the propagation, i.e. Shadowing
\code{logDistance} uses the following inputs:
}
\description{
Produces the estiamted power in each measured point based on the log Distance path loss model
}
\seealso{
Other path loss models: \code{\link{cost231}},
  \code{\link{erceg}}, \code{\link{okumura}},
  \code{\link{walfish}}
}
\concept{path loss models}
