% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/okumura.R
\name{okumura}
\alias{okumura}
\title{Okumura - Hata path loss model}
\usage{
okumura(Data, freq, typeOfArea)
}
\arguments{
\item{Data}{A dataframe with columns the distance from the transmiter and the measured power at that point.}

\item{freq}{the frequency in MHz of the transmition}

\item{typeOfArea}{the type of area the measurement took place, i.e.: Urban or Suburban}
}
\value{
A list with the estimated Intercept of the model and the residuals of the propagation, i.e. Shadowing
\code{okumura} uses the following inputs:
}
\description{
Produces the estiamted power in each measured point based on the Okumura-Hata path loss model
}
\seealso{
Other path loss models: \code{\link{cost231}},
  \code{\link{erceg}}, \code{\link{logDistance}},
  \code{\link{walfish}}
}
\concept{path loss models}
