% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walfish.R
\name{walfish}
\alias{walfish}
\title{Walfish-Ikegami path loss model}
\usage{
walfish(Data, freq)
}
\arguments{
\item{Data}{A dataframe with columns the distance from the transmiter and the measured power at that point.}

\item{freq}{the frequency in MHz of the transmition}
}
\value{
A list with the estimated Intercept of the model and the residuals of the propagation, i.e. Shadowing
\code{walfish} uses the following inputs:
}
\description{
Produces the estiamted power in each measured point based on the Walfish-Ikegami path loss model
}
\seealso{
Other path loss models: \code{\link{cost231}},
  \code{\link{erceg}}, \code{\link{logDistance}},
  \code{\link{okumura}}
}
\concept{path loss models}
